Fixed the frontend

+++ backend/controller/goalController.js
@@ -1,8 +1,10 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalMoldel')
+const User = require('../models/userModel')
+
 
 const getGoals = asyncHandler(async(req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user : req.user.id})
     res.json(goals)
 })
 
@@ -12,7 +14,8 @@ const setGoal = asyncHandler(async(req, res) => {
         throw new Error('Please add a textfield')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -23,6 +26,18 @@ const updateGoals = asyncHandler(async(req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+    //check for user
+    const user = await User.findById(req.user.id)
+    if (!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('user not authorised')
     }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
@@ -37,6 +52,19 @@ const deleteGoal = asyncHandler(async(req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    const user = await User.findById(req.user.id)
+    if (!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    //make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('user not authorised')
+    }
+    
     await goal.deleteOne({_id : req.params.id})
 
     res.json({ id:req.params.id })
